[types] Add @material-ui/types package
