+++ packages/material-ui-styles/package.json
@@ -39,6 +39,7 @@
   "dependencies": {
     "@babel/runtime": "^7.2.0",
     "@emotion/hash": "^0.7.1",
+    "@material-ui/types": "^1.0.0",
     "@material-ui/utils": "^4.0.0-beta.0",
     "clsx": "^1.0.2",
     "deepmerge": "^3.0.0",
