+++ packages/material-ui-types/README.md
@@ -1,0 +1,1 @@
+# @material-ui/types
