+++ packages/material-ui-types/index.d.ts
@@ -1,0 +1,44 @@
+// disable automatic export
+export {};
+
+/**
+ * `T extends ConsistentWith<T, U>` means that where `T` has overlapping properties with
+ * `U`, their value types do not conflict.
+ *
+ * @internal
+ */
+export type ConsistentWith<DecorationTargetProps, InjectedProps> = {
+  [P in keyof DecorationTargetProps]: P extends keyof InjectedProps
+    ? InjectedProps[P] extends DecorationTargetProps[P]
+      ? DecorationTargetProps[P]
+      : InjectedProps[P]
+    : DecorationTargetProps[P]
+};
+
+/**
+ * a function that takes {component} and returns a component that passes along
+ * all the props to {component} except the {InjectedProps} and will accept
+ * additional {AdditionalProps}
+ */
+export type PropInjector<InjectedProps, AdditionalProps = {}> = <
+  C extends React.ComponentType<ConsistentWith<React.ComponentProps<C>, InjectedProps>>
+>(
+  component: C,
+) => React.ComponentType<
+  Omit<JSX.LibraryManagedAttributes<C, React.ComponentProps<C>>, keyof InjectedProps> &
+    AdditionalProps
+>;
+
+/**
+ * Remove properties `K` from `T`.
+ *
+ * @internal
+ */
+export type Omit<T, K extends keyof any> = T extends any ? Pick<T, Exclude<keyof T, K>> : never;
+
+/**
+ * Like `T & U`, but using the value types from `U` where their properties overlap.
+ *
+ * @internal
+ */
+export type Overwrite<T, U> = Omit<T, keyof U> & U;
