+++ packages/material-ui-types/package.json
@@ -1,0 +1,33 @@
+{
+  "name": "@material-ui/types",
+  "version": "1.0.0",
+  "private": false,
+  "author": "Material-UI Team",
+  "description": "Material-UI Types - Utility types for Material-UI.",
+  "types": "./index.d.ts",
+  "keywords": [
+    "react",
+    "react-component",
+    "material design",
+    "material-ui",
+    "types"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/mui-org/material-ui.git"
+  },
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/mui-org/material-ui/issues"
+  },
+  "homepage": "https://github.com/mui-org/material-ui/tree/next/packages/material-ui-types",
+  "scripts": {
+    "release": "npm publish --tag next",
+    "test": "echo 'No runtime test. Type tests are run with the `typescript` script.'",
+    "typescript": "tslint -p tsconfig.json \"*.{ts,tsx}\""
+  },
+  "sideEffects": false,
+  "publishConfig": {
+    "access": "public"
+  }
+}
