+++ packages/material-ui/package.json
@@ -39,6 +39,7 @@
     "@babel/runtime": "^7.2.0",
     "@material-ui/styles": "^4.0.0-beta.0",
     "@material-ui/system": "^4.0.0-beta.0",
+    "@material-ui/types": "^1.0.0",
     "@material-ui/utils": "^4.0.0-beta.0",
     "@types/react-transition-group": "^2.0.16",
     "clsx": "^1.0.2",
