+++ packages/material-ui/src/Grow/Grow.d.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { Omit } from '..';
+import { Omit } from '@material-ui/types';
 import { Theme } from '../styles/createMuiTheme';
 import { TransitionProps } from '../transitions/transition';
