+++ packages/material-ui/src/Link/Link.d.ts
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { OverridableComponent, SimplifiedPropsOf } from '../OverridableComponent';
-import { Omit } from '..';
+import { Omit } from '@material-ui/types';
 import { TypographyProps } from '../Typography';
 
 declare const Link: OverridableComponent<{
