+++ packages/material-ui/src/index.d.ts
@@ -1,6 +1,13 @@
 import * as React from 'react';
+import { ConsistentWith, Omit } from '@material-ui/types';
 import { StyledComponentProps } from './styles';
 export { StyledComponentProps };
+
+/**
+ * @deprecated
+ * Import from `@material-ui/types` instead
+ */
+export { Omit };
 
 /**
  * All standard components exposed by `material-ui` are `StyledComponents` with
@@ -41,48 +48,6 @@ export interface Color {
   A400: string;
   A700: string;
 }
-
-/**
- * Remove properties `K` from `T`.
- *
- * @internal
- */
-export type Omit<T, K extends keyof any> = T extends any ? Pick<T, Exclude<keyof T, K>> : never;
-
-/**
- * `T extends ConsistentWith<T, U>` means that where `T` has overlapping properties with
- * `U`, their value types do not conflict.
- *
- * @internal
- */
-export type ConsistentWith<DecorationTargetProps, InjectedProps> = {
-  [P in keyof DecorationTargetProps]: P extends keyof InjectedProps
-    ? InjectedProps[P] extends DecorationTargetProps[P]
-      ? DecorationTargetProps[P]
-      : InjectedProps[P]
-    : DecorationTargetProps[P]
-};
-
-/**
- * a function that takes {component} and returns a component that passes along
- * all the props to {component} except the {InjectedProps} and will accept
- * additional {AdditionalProps}
- */
-export type PropInjector<InjectedProps, AdditionalProps = {}> = <
-  C extends React.ComponentType<ConsistentWith<React.ComponentProps<C>, InjectedProps>>
->(
-  component: C,
-) => React.ComponentType<
-  Omit<JSX.LibraryManagedAttributes<C, React.ComponentProps<C>>, keyof InjectedProps> &
-    AdditionalProps
->;
-
-/**
- * Like `T & U`, but using the value types from `U` where their properties overlap.
- *
- * @internal
- */
-export type Overwrite<T, U> = Omit<T, keyof U> & U;
 
 export namespace PropTypes {
   type Alignment = 'inherit' | 'left' | 'center' | 'right' | 'justify';
